requirements.txt
flask
boto3
mysql-connector